feat(builder): add min length validation option for password fields
